datetime done

+++ src/main.rs
@@ -4,5 +4,5 @@ use floton::datetime;
 fn main() {
     println!("---- Floton DB ----");
     println!("TID: {}", tlocal::tid());
-    println!("Unix Time: {}", datetime::unix_time());
+    println!("Unix Time: {}", datetime::unix_time() as u64);
 }
